+++ content/api/serializers.py
@@ -32,8 +32,7 @@ from content.models import (
     Review,
     TargetedFundraising,
     TypeDocument,
-
-from .validators import validate_phone_number
+)
 
 class GratitudeSerializer(serializers.ModelSerializer):
     """Сериализатор для модели Gratitude."""
@@ -280,11 +279,3 @@ class EmployeeDetailSerializer(serializers.ModelSerializer):
         return CategorySerializer(
             categories, many=True, context=self.context
         ).data
-
-
-class FeedbackFormSerializer(serializers.Serializer):
-    """Сериализатор для формы обратной связи."""
-
-    name = serializers.CharField(max_length=255)
-    phone_number = serializers.CharField(max_length=25,validators=[validate_phone_number])
-    message = serializers.CharField(max_length=1000) # около 200 слов
