+++ content/api/urls.py
@@ -35,6 +35,6 @@ v1_router_api.register(
 api_urls.extend(v1_router_api.urls)
 
 urlpatterns = [
-    path('v1/feedback/', views.FeedbackFormView.as_view(), name='fitbak_form'),
+    path('v1/form/', include('forms.urls')),
     path('v1/', include(api_urls)),
 ]
